Generaliser hvordan man setter opp aktiviteter for ReiseTilSamling og PassAvBarn
