+++ src/frontend/barnetilsyn/tekster/aktivitet.ts
@@ -1,68 +1,21 @@
-import { tekstArbeidsrettedeAktiviteter } from '../../tekster/aktivitet';
-import { JaNeiTilTekst } from '../../tekster/felles';
+import {
+    AktivitetInnhold,
+    aktivitetTeksterFelles,
+    HvilkenAktivitet,
+    plukkAktivitetTekster,
+    tekstArbeidsrettedeAktiviteter,
+} from '../../tekster/aktivitet';
 import { AnnenAktivitetType } from '../../typer/aktivitet';
-import { JaNei } from '../../typer/søknad';
-import { InlineLenke, LesMer, Radiogruppe, TekstElement } from '../../typer/tekst';
 
-interface AktivitetInnhold {
-    guide_innhold: TekstElement<string[]>;
-    radio_lønnet_tiltak: Radiogruppe<JaNei>;
-    checkbox_velge_aktivitet_feilmelding: TekstElement<string>;
-    radio_annet: Radiogruppe<AnnenAktivitetType>;
-    radio_annet_uten_registeraktivitet: Radiogruppe<AnnenAktivitetType>;
-    radio_annet_lesmer: LesMer<InlineLenke>;
-    radio_annet_lesmer_hva_betyr_alternativene: HvaBetyrAlternativene;
-    radio_annet_feilmelding: TekstElement<string>;
-    ingen_registrerte_aktiviterer_overskrift: TekstElement<string>;
-    tittel: TekstElement<string>;
-    ingen_aktivitet_infoalert_title: TekstElement<string>;
-    ingen_aktivitet_infoalert_innhold: IngenAktivitet;
-    lønnet_tiltak_infoalert_innhold: TekstElement<string[]>;
-    radio_lønnet_tiltak_feilmelding: TekstElement<string>;
-    radio_fortsatt_søke: Radiogruppe<JaNei>;
-    søker_fra_label: TekstElement<string>;
-    søker_fra_lesmer: LesMer<string[]>;
-    søker_fra_dato_feilmelding: TekstElement<string>;
-    hvilken_aktivitet: HvilkenAktivitet;
-}
-
-interface HvilkenAktivitet {
-    spm: TekstElement<string>;
-    les_mer: {
-        header: TekstElement<string>;
-        header_ingen_registrerte_aktiviteter: TekstElement<string>;
-        del1: TekstElement<string[]>;
-        del2_lenker: TekstElement<InlineLenke>[];
-        del3: TekstElement<InlineLenke>;
-    };
-}
-interface HvaBetyrAlternativene {
-    header: TekstElement<string>;
-    innhold: {
-        del1: TekstElement<string[]>;
-        del2_lenker: TekstElement<InlineLenke>[];
-    };
-}
-interface IngenAktivitet {
-    del1: TekstElement<string[]>;
-    del2_lenker: TekstElement<InlineLenke>[];
-}
-
-export const AktivitetTypeTilTekst: Record<AnnenAktivitetType, TekstElement<string>> = {
-    TILTAK: { nb: 'Tiltak / arbeidsrettet utredning' },
-    UTDANNING: {
-        nb: 'Utdanning godkjent av Nav',
-    },
-    ARBEIDSSØKER: {
-        nb: 'Jeg er arbeidssøker',
-    },
-    INGEN_AKTIVITET: {
-        nb: 'Har ingen arbeidsrettet aktivitet',
-    },
-};
+export const AktivitetTypeTilTekstPassAvBarn = plukkAktivitetTekster(
+    AnnenAktivitetType.TILTAK,
+    AnnenAktivitetType.UTDANNING,
+    AnnenAktivitetType.ARBEIDSSØKER,
+    AnnenAktivitetType.INGEN_AKTIVITET
+);
 
 const hvilkenAktivitet: HvilkenAktivitet = {
-    spm: { nb: 'Hvilken aktivitet søker du om støtte i forbindelse med?' },
+    spm: tekstArbeidsrettedeAktiviteter.spm,
     les_mer: {
         header: tekstArbeidsrettedeAktiviteter.lesMer.header,
         header_ingen_registrerte_aktiviteter:
@@ -111,46 +64,8 @@ const hvilkenAktivitet: HvilkenAktivitet = {
 };
 
 export const aktivitetTekster: AktivitetInnhold = {
-    søker_fra_lesmer: {
-        header: { nb: 'Hvilken dato velger jeg?' },
-        innhold: {
-            nb: [
-                'Det vanligste er å velge datoen utdanningen din starter eller fra da du ble registrert arbeidssøker. Du får bare dekket utgifter så lenge du har en arbeidsrettet aktivitet. ',
-                'Du kan søke tilbake i tid, men som hovedregel kan vi bare innvilge inntil 3 måneder fra datoen du søker.',
-            ],
-        },
-    },
-    søker_fra_label: { nb: 'Fra hvilken dato søker du om støtte til pass?' },
-    søker_fra_dato_feilmelding: { nb: 'Du må fylle inn en gyldig dato' },
-    tittel: {
-        nb: 'Arbeidsrettet aktivitet',
-    },
-    guide_innhold: {
-        nb: [
-            'For å få denne stønaden må du delta på tiltak, ta en utdannelse godkjent av Nav, være arbeidssøker eller gjennomføre en arbeidsrettet utredning.',
-            'Vi viser aktivitetene som er registrert på deg de siste 3 månedene.',
-        ],
-    },
+    ...aktivitetTeksterFelles,
     hvilken_aktivitet: hvilkenAktivitet,
-    radio_lønnet_tiltak: {
-        header: {
-            nb: 'Mottar du lønn gjennom et tiltak?',
-        },
-        alternativer: JaNeiTilTekst,
-    },
-    checkbox_velge_aktivitet_feilmelding: {
-        nb: 'Du må svare på hvilken aktivitet du søker om støtte i forbindelse med.',
-    },
-    lønnet_tiltak_infoalert_innhold: {
-        nb: [
-            'Hvis du mottar lønn i tiltaket kan du fortsatt søke, men det kan hende du får avslag.',
-            'Som lærling kan du få støtte til pass av barn, selv om du mottar lønn.',
-        ],
-    },
-    radio_lønnet_tiltak_feilmelding: {
-        nb: 'Du må svare på om du mottar lønn gjennom et tiltak.',
-    },
-    ingen_aktivitet_infoalert_title: { nb: 'Ingen arbeidsrettet aktivitet?' },
     ingen_aktivitet_infoalert_innhold: {
         del1: {
             nb: [
@@ -181,25 +96,6 @@ export const aktivitetTekster: AktivitetInnhold = {
                 ],
             },
         ],
-    },
-    radio_fortsatt_søke: {
-        header: {
-            nb: 'Vil du fortsatt søke nå?',
-        },
-        alternativer: JaNeiTilTekst,
-    },
-    radio_annet_lesmer: {
-        header: { nb: 'Søke lengre tilbake enn 3 måneder?' },
-        innhold: {
-            nb: [
-                'Du må sende inn søknad på papir hvis du skal søke om en aktivitet som ble avsluttet for mer enn 3 måneder siden. ',
-                {
-                    tekst: 'Søknad på papir',
-                    url: 'https://www.nav.no/fyllut/nav111215b?sub=paper',
-                },
-                '.',
-            ],
-        },
     },
     radio_annet_lesmer_hva_betyr_alternativene: {
         header: { nb: 'Hva betyr alternativene?' },
@@ -239,16 +135,10 @@ export const aktivitetTekster: AktivitetInnhold = {
     },
     radio_annet: {
         header: { nb: 'Hvilken annen type arbeidsrettet aktivitet har du?' },
-        alternativer: AktivitetTypeTilTekst,
+        alternativer: AktivitetTypeTilTekstPassAvBarn,
     },
     radio_annet_uten_registeraktivitet: {
         header: { nb: 'Hvilken arbeidsrettet aktivitet har du?' },
-        alternativer: AktivitetTypeTilTekst,
-    },
-    radio_annet_feilmelding: {
-        nb: 'Du må svare på hvilken aktivitet du søker om støtte i forbindelse med.',
-    },
-    ingen_registrerte_aktiviterer_overskrift: {
-        nb: 'Vi fant dessverre ingen arbeidsrettede aktiviteter som er registrert på deg.',
+        alternativer: AktivitetTypeTilTekstPassAvBarn,
     },
 };
